+++ packages/testland/e2e/testcases/death/declaration/death-declaration-1.spec.ts
@@ -58,7 +58,7 @@ test.describe.serial('1. Death declaration case - 1', () => {
       date: getRandomDate(0, 20),
       causeOfDeathEstablished: true,
       sourceCauseDeath: 'Physician',
-      placeOfDeath: "Deceased's usual place of residence"
+      placeOfDeath: 'Other'
     },
     causeOfDeathDetails: {
       causeOfDeathA: {